Added pings to detect failed websockets ; v1.2
Grouped all send logic in a broadcast_and_clean function.
Send a message to all, and remove all failed websockets.

This function is then used in both connect_clients (for cleaning) and
notify_clients (for sending the message).

For cleaning, an empty string JSON object is sent (dummy value).
-> Added filtering in clients to not print an error for this object.

This allows earlier detection of failed websockets, even if no chat
message is sent.
In particular, the pattern of reloading the page repeateadly (F5 F5
F5...) does not create infinite failed websockets.

With all of that, go to v1.2.

---

I temporarily considered maintaining a user list instead of a basic
ping.
However the current design separates posting of message (AJAX on post
end points) and notification, so it is impossible to know which user is
behind which socket.
Without moving to a better framework, this cannot be done cleanly, and
won't be done.